change fix id type from int to long

+++ Badcamp/Badcamp.Application/UseCases/EventCases/GetEventByArtistRequest.cs
@@ -9,6 +9,6 @@ namespace Badcamp.Application.UseCases.EventCases
 {
     public class GetEventByArtistRequest : IRequest<Response>
     {
-        public int ArtistId { get; set; }
+        public long ArtistId { get; set; }
     }
 }
